Start on main scene
- Main game scene
- Menu, level gen, player movement
- Features interact together
