+++ Assets/Resources/LevelGen/TestSet/Tiles/1,Blue.prefab
@@ -12,7 +12,7 @@ GameObject:
   - component: {fileID: 7551444753501090845}
   - component: {fileID: 1636686708471016859}
   - component: {fileID: 4843447589994901004}
-  m_Layer: 0
+  m_Layer: 6
   m_Name: 1,Blue
   m_TagString: Untagged
   m_Icon: {fileID: 0}
