+++ Assets/Scenes/Main.unity.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: e3213084b7a1cd84a98e8b4088d7a9cf
-folderAsset: yes
+guid: 543d3f7ec696dd644be0b43707af7d3d
 DefaultImporter:
   externalObjects: {}
   userData:
