+++ Assets/Scenes/Testing/Player.unity
@@ -75,18 +75,18 @@ LightmapSettings:
     m_BakeBackend: 1
     m_PVRSampling: 1
     m_PVRDirectSampleCount: 32
-    m_PVRSampleCount: 512
+    m_PVRSampleCount: 500
     m_PVRBounces: 2
-    m_PVREnvironmentSampleCount: 256
+    m_PVREnvironmentSampleCount: 500
     m_PVREnvironmentReferencePointCount: 2048
-    m_PVRFilteringMode: 1
-    m_PVRDenoiserTypeDirect: 1
-    m_PVRDenoiserTypeIndirect: 1
-    m_PVRDenoiserTypeAO: 1
+    m_PVRFilteringMode: 2
+    m_PVRDenoiserTypeDirect: 0
+    m_PVRDenoiserTypeIndirect: 0
+    m_PVRDenoiserTypeAO: 0
     m_PVRFilterTypeDirect: 0
     m_PVRFilterTypeIndirect: 0
     m_PVRFilterTypeAO: 0
-    m_PVREnvironmentMIS: 1
+    m_PVREnvironmentMIS: 0
     m_PVRCulling: 1
     m_PVRFilteringGaussRadiusDirect: 1
     m_PVRFilteringGaussRadiusIndirect: 5
@@ -123,64 +123,64 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
---- !u!1001 &8327526105342537417
+--- !u!1001 &213757027181651902
 PrefabInstance:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 1117644708746057418, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
-      propertyPath: m_RootOrder
-      value: 2
-      objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380156, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
-      propertyPath: m_Name
-      value: SpawnTile
-      objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380159, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+    - target: {fileID: 213757026412541186, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
       propertyPath: m_RootOrder
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380159, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+    - target: {fileID: 213757026412541186, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
       propertyPath: m_LocalPosition.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380159, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+    - target: {fileID: 213757026412541186, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
       propertyPath: m_LocalPosition.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380159, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+    - target: {fileID: 213757026412541186, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
       propertyPath: m_LocalPosition.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380159, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+    - target: {fileID: 213757026412541186, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
       propertyPath: m_LocalRotation.w
       value: 1
       objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380159, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+    - target: {fileID: 213757026412541186, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
       propertyPath: m_LocalRotation.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380159, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+    - target: {fileID: 213757026412541186, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
       propertyPath: m_LocalRotation.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380159, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+    - target: {fileID: 213757026412541186, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
       propertyPath: m_LocalRotation.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380159, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+    - target: {fileID: 213757026412541186, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
       propertyPath: m_LocalEulerAnglesHint.x
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380159, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+    - target: {fileID: 213757026412541186, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
       propertyPath: m_LocalEulerAnglesHint.y
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 8327526105932380159, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+    - target: {fileID: 213757026412541186, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 213757026412541189, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
+      propertyPath: webFunctions
+      value: 
+      objectReference: {fileID: 0}
+    - target: {fileID: 213757026412541193, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
+      propertyPath: m_Name
+      value: Player
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 74bddf4668f957f48bf9c7af8d167558, type: 3}
+  m_SourcePrefab: {fileID: 100100000, guid: 3e99a6cabf8294446b906b81fbdb3993, type: 3}
