+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,15 +6,15 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/Menus.unity
+    path: Assets/Scenes/Testing/Menus.unity
     guid: 4e042f8929f77cb4b8646298128b0aee
   - enabled: 1
-    path: Assets/Scenes/Player.unity
+    path: Assets/Scenes/Testing/Player.unity
     guid: 39de4bafc8fd6c04eabac44a86e06c11
   - enabled: 1
     path: Assets/Scenes/Spawn.unity
     guid: f9a1b42dc369d3144ae46286c0cf56e2
   - enabled: 1
-    path: Assets/Scenes/TestLevel.unity
+    path: Assets/Scenes/Testing/Level Testing.unity
     guid: 19f79f195ce8b484fb7df3941d6b4381
   m_configObjects: {}
